Refactor slicing utilities out of the Content class into a hidden submodule.
