+++ src/awkward/_v2/contents/content.py
@@ -8,6 +8,7 @@ except ImportError:
     from collections import Iterable
 
 import awkward as ak
+from awkward._v2._slicing import NestedIndexError
 from awkward._v2.tmp_for_testing import v1_to_v2, v2_to_v1
 
 np = ak.nplike.NumpyMetadata.instance()
@@ -90,14 +91,8 @@ class Content(object):
                 else:
                     raise NestedIndexError(self, slicer, message)
 
-    def _headtail(self, oldtail):
-        if len(oldtail) == 0:
-            return (), ()
-        else:
-            return oldtail[0], oldtail[1:]
-
     def _getitem_next_field(self, head, tail, advanced):
-        nexthead, nexttail = self._headtail(tail)
+        nexthead, nexttail = ak._v2._slicing.headtail(tail)
         return self._getitem_field(head)._getitem_next(nexthead, nexttail, advanced)
 
     def _getitem_next_fields(self, head, tail, advanced):
@@ -107,13 +102,13 @@ class Content(object):
                 only_fields.append(x)
             else:
                 not_fields.append(x)
-        nexthead, nexttail = self._headtail(tuple(not_fields))
+        nexthead, nexttail = ak._v2._slicing.headtail(tuple(not_fields))
         return self._getitem_fields(head, tuple(only_fields))._getitem_next(
             nexthead, nexttail, advanced
         )
 
     def _getitem_next_newaxis(self, tail, advanced):
-        nexthead, nexttail = self._headtail(tail)
+        nexthead, nexttail = ak._v2._slicing.headtail(tail)
         return ak._v2.contents.RegularArray(
             self._getitem_next(nexthead, nexttail, advanced),
             1,  # size
@@ -130,7 +125,7 @@ class Content(object):
         )
 
         if len(tail) == 0 or mindepth - 1 == maxdepth - 1 == dimlength:
-            nexthead, nexttail = self._headtail(tail)
+            nexthead, nexttail = ak._v2._slicing.headtail(tail)
             return self._getitem_next(nexthead, nexttail, advanced)
 
         elif mindepth - 1 == dimlength or maxdepth - 1 == dimlength:
@@ -142,65 +137,6 @@ class Content(object):
 
         else:
             return self._getitem_next(slice(None), (Ellipsis,) + tail, advanced)
-
-    def _getitem_broadcast(self, items, nplike):
-        lookup = []
-        broadcastable = []
-        for item in items:
-            if (
-                isinstance(
-                    item,
-                    (
-                        slice,
-                        list,
-                        ak._v2.contents.ListOffsetArray,
-                        ak._v2.contents.IndexedOptionArray,
-                    ),
-                )
-                or ak._util.isstr(item)
-                or item is np.newaxis
-                or item is Ellipsis
-            ):
-                lookup.append(None)
-            elif isinstance(item, int):
-                lookup.append(len(broadcastable))
-                broadcastable.append(item)
-            else:
-                lookup.append(len(broadcastable))
-                broadcastable.append(nplike.asarray(item))
-
-        broadcasted = nplike.broadcast_arrays(*broadcastable)
-
-        out = []
-        for i, item in zip(lookup, items):
-            if i is None:
-                out.append(item)
-            else:
-                x = broadcasted[i]
-                if len(x.shape) == 0:
-                    out.append(int(x))
-                else:
-                    if issubclass(x.dtype.type, np.int64):
-                        out.append(ak._v2.index.Index64(x.reshape(-1)))
-                        out[-1].metadata["shape"] = x.shape
-                    elif issubclass(x.dtype.type, np.integer):
-                        out.append(ak._v2.index.Index64(x.astype(np.int64).reshape(-1)))
-                        out[-1].metadata["shape"] = x.shape
-                    elif issubclass(x.dtype.type, (np.bool_, bool)):
-                        if len(x.shape) == 1:
-                            out.append(ak._v2.index.Index64(nplike.nonzero(x)[0]))
-                            out[-1].metadata["shape"] = x.shape
-                        else:
-                            for w in nplike.nonzero(x):
-                                out.append(ak._v2.index.Index64(w))
-                    else:
-                        raise TypeError(
-                            "array slice must be an array of integers or booleans, not\n\n    {0}".format(
-                                repr(x).replace("\n", "\n    ")
-                            )
-                        )
-
-        return tuple(out)
 
     def _getitem_next_regular_missing(self, head, tail, advanced, raw, length):
         # if this is in a tuple-slice and really should be 0, it will be trimmed later
@@ -352,8 +288,8 @@ class Content(object):
                 if len(where) == 0:
                     return self
 
-                nextwhere = self._getitem_broadcast(
-                    [self._prepare_tuple_item(x) for x in where],
+                nextwhere = ak._v2._slicing.getitem_broadcast(
+                    [ak._v2._slicing.prepare_tuple_item(x) for x in where],
                     self.nplike,
                 )
 
@@ -397,7 +333,7 @@ class Content(object):
                         )
                     )
 
-                out = self._getitem_next_array_wrap(
+                out = ak._v2._slicing.getitem_next_array_wrap(
                     self._carry(carry, allow_lazy, NestedIndexError), where.shape
                 )
                 if len(out) == 0:
@@ -474,307 +410,6 @@ at inner {2} of length {3}, using sub-slice {4}.{5}""".format(
                     ),
                 )
             )
-
-    def _prepare_tuple_item(self, item):
-        if ak._util.isint(item):
-            return int(item)
-
-        elif isinstance(item, slice):
-            return item
-
-        elif ak._util.isstr(item):
-            return item
-
-        elif item is np.newaxis:
-            return item
-
-        elif item is Ellipsis:
-            return item
-
-        elif isinstance(item, ak.highlevel.Array):
-            return self._prepare_tuple_item(item.layout)
-
-        elif isinstance(item, ak.layout.Content):
-            return self._prepare_tuple_item(v1_to_v2(item))
-
-        elif isinstance(item, ak._v2.contents.EmptyArray):
-            return self._prepare_tuple_item(item.toNumpyArray(np.int64))
-
-        elif isinstance(item, ak._v2.contents.NumpyArray):
-            return item.data
-
-        elif isinstance(item, ak._v2.contents.Content):
-            return self._prepare_tuple_bool_to_int(self._prepare_tuple_nested(item))
-
-        elif isinstance(item, Iterable) and all(ak._util.isstr(x) for x in item):
-            return list(item)
-
-        elif isinstance(item, Iterable):
-            layout = v1_to_v2(ak.operations.convert.to_layout(item))
-            as_nplike = layout.maybe_to_nplike(layout.nplike)
-            if as_nplike is None:
-                return self._prepare_tuple_item(layout)
-            else:
-                return self._prepare_tuple_item(ak._v2.contents.NumpyArray(as_nplike))
-
-        else:
-            raise TypeError(
-                "only integers, slices (`:`), ellipsis (`...`), np.newaxis (`None`), "
-                "integer/boolean arrays (possibly with variable-length nested "
-                "lists or missing values), field name (str) or names (non-tuple "
-                "iterable of str) are valid indices for slicing, not\n\n    "
-                + repr(item).replace("\n", "\n    ")
-            )
-
-    def _prepare_tuple_RegularArray_toListOffsetArray64(self, item):
-        if isinstance(item, ak._v2.contents.RegularArray):
-
-            next = item.toListOffsetArray64()
-            return ak._v2.contents.ListOffsetArray(
-                next.offsets,
-                self._prepare_tuple_RegularArray_toListOffsetArray64(next.content),
-                identifier=item.identifier,
-                parameters=item.parameters,
-            )
-
-        elif isinstance(item, ak._v2.contents.NumpyArray):
-            return item
-
-        else:
-            raise AssertionError(type(item))
-
-    def _prepare_tuple_nested(self, item):
-        if isinstance(item, ak._v2.contents.EmptyArray):
-            return self._prepare_tuple_nested(item.toNumpyArray(np.int64))
-
-        elif isinstance(item, ak._v2.contents.NumpyArray) and issubclass(
-            item.dtype.type, (bool, np.bool_, np.integer)
-        ):
-            if issubclass(item.dtype.type, (bool, np.bool_, np.int64)):
-                next = item
-            else:
-                next = ak._v2.contents.NumpyArray(
-                    item.data.astype(np.int64),
-                    identifier=item.identifier,
-                    parameters=item.parameters,
-                    nplike=item.nplike,
-                )
-            next = next.toRegularArray()
-            next = self._prepare_tuple_RegularArray_toListOffsetArray64(next)
-            return next
-
-        elif (
-            isinstance(
-                item,
-                ak._v2.contents.ListOffsetArray,
-            )
-            and issubclass(item.offsets.dtype.type, np.int64)
-        ):
-            return ak._v2.contents.ListOffsetArray(
-                item.offsets,
-                self._prepare_tuple_nested(item.content),
-                identifier=item.identifier,
-                parameters=item.parameters,
-            )
-
-        elif isinstance(
-            item,
-            (
-                ak._v2.contents.ListOffsetArray,
-                ak._v2.contents.ListArray,
-                ak._v2.contents.RegularArray,
-            ),
-        ):
-            next = item.toListOffsetArray64(False)
-            return self._prepare_tuple_nested(next)
-
-        elif isinstance(
-            item,
-            ak._v2.contents.IndexedArray,
-        ):
-            next = item.project()
-            return self._prepare_tuple_nested(next)
-
-        elif (
-            isinstance(
-                item,
-                ak._v2.contents.IndexedOptionArray,
-            )
-            and issubclass(item.index.dtype.type, np.int64)
-        ):
-            return ak._v2.contents.IndexedOptionArray(
-                item.index,
-                self._prepare_tuple_nested(item.content),
-                identifier=item.identifier,
-                parameters=item.parameters,
-            )._simplify_optiontype()
-
-        elif isinstance(
-            item,
-            (
-                ak._v2.contents.IndexedOptionArray,
-                ak._v2.contents.ByteMaskedArray,
-                ak._v2.contents.BitMaskedArray,
-                ak._v2.contents.UnmaskedArray,
-            ),
-        ):
-            next = item.toIndexedOptionArray64()
-            return self._prepare_tuple_nested(next)
-
-        else:
-            raise TypeError(
-                "only integers, slices (`:`), ellipsis (`...`), np.newaxis (`None`), "
-                "integer/boolean arrays (possibly with variable-length nested "
-                "lists or missing values), field name (str) or names (non-tuple "
-                "iterable of str) are valid indices for slicing, not\n\n    "
-                + repr(item).replace("\n", "\n    ")
-            )
-
-    def _prepare_tuple_bool_to_int(self, item):
-        if (
-            isinstance(item, ak._v2.contents.ListOffsetArray)
-            and isinstance(item.content, ak._v2.contents.NumpyArray)
-            and issubclass(item.content.dtype.type, (bool, np.bool_))
-        ):
-            localindex = item.localindex(axis=1)
-            nextcontent = localindex.content.data[item.content.data]
-
-            cumsum = item.nplike.empty(len(item.content.data) + 1, np.int64)
-            cumsum[0] = 0
-            cumsum[1:] = item.nplike.cumsum(item.content.data)
-            nextoffsets = cumsum[item.offsets]
-
-            return ak._v2.contents.ListOffsetArray(
-                ak._v2.index.Index64(nextoffsets),
-                ak._v2.contents.NumpyArray(nextcontent, nplike=item.nplike),
-            )
-
-        elif (
-            isinstance(item, ak._v2.contents.ListOffsetArray)
-            and isinstance(item.content, ak._v2.contents.IndexedOptionArray)
-            and isinstance(item.content.content, ak._v2.contents.NumpyArray)
-            and issubclass(item.content.content.dtype.type, (bool, np.bool_))
-        ):
-            # missing values as any integer other than -1 are extremely rare
-            isnegative = item.content.index.data < 0
-            if item.nplike.any(item.content.index.data < -1):
-                safeindex = item.content.index.data.copy()
-                safeindex[isnegative] = -1
-            else:
-                safeindex = item.content.index.data
-
-            # expanded is a new buffer (can be modified in-place)
-            if len(item.content.content.data) > 0:
-                expanded = item.content.content.data[safeindex]
-            else:
-                expanded = item.content.content.data.nplike.ones(
-                    len(safeindex), np.bool_
-                )
-
-            localindex = item.localindex(axis=1)
-
-            # nextcontent does not include missing values
-            expanded[isnegative] = False
-            nextcontent = localindex.content.data[expanded]
-
-            # list offsets do include missing values
-            expanded[isnegative] = True
-            cumsum = item.nplike.empty(len(expanded) + 1, np.int64)
-            cumsum[0] = 0
-            cumsum[1:] = item.nplike.cumsum(expanded)
-            nextoffsets = cumsum[item.offsets]
-
-            # outindex fits into the lists; non-missing are sequential
-            outindex = item.nplike.full(nextoffsets[-1], -1, np.int64)
-            outindex[~isnegative[expanded]] = item.nplike.arange(
-                len(nextcontent), dtype=np.int64
-            )
-
-            return ak._v2.contents.ListOffsetArray(
-                ak._v2.index.Index64(nextoffsets),
-                ak._v2.contents.IndexedOptionArray(
-                    ak._v2.index.Index(outindex),
-                    ak._v2.contents.NumpyArray(nextcontent, nplike=item.nplike),
-                ),
-            )
-
-        elif isinstance(item, ak._v2.contents.ListOffsetArray):
-            return ak._v2.contents.ListOffsetArray(
-                item.offsets,
-                item.content,
-                identifier=item.identifier,
-                parameters=item.parameters,
-            )
-
-        elif isinstance(item, ak._v2.contents.IndexedOptionArray):
-            if isinstance(item.content, ak._v2.contents.ListOffsetArray):
-                content = self._prepare_tuple_bool_to_int(item.content)
-                return ak._v2.contents.IndexedOptionArray(
-                    item.index,
-                    content,
-                    identifier=item.identifier,
-                    parameters=item.parameters,
-                )
-
-            if isinstance(item.content, ak._v2.contents.NumpyArray) and issubclass(
-                item.content.dtype.type, (bool, np.bool_)
-            ):
-                # missing values as any integer other than -1 are extremely rare
-                isnegative = item.index.data < 0
-                if item.nplike.any(item.index.data < -1):
-                    safeindex = item.index.data.copy()
-                    safeindex[isnegative] = -1
-                else:
-                    safeindex = item.index.data
-
-                # expanded is a new buffer (can be modified in-place)
-                if len(item.content.data) > 0:
-                    expanded = item.content.data[safeindex]
-                else:
-                    expanded = item.content.data.nplike.ones(len(safeindex), np.bool_)
-
-                # nextcontent does not include missing values
-                expanded[isnegative] = False
-                nextcontent = item.nplike.nonzero(expanded)[0]
-
-                # outindex does include missing values
-                expanded[isnegative] = True
-                lenoutindex = item.nplike.count_nonzero(expanded)
-
-                # non-missing are sequential
-                outindex = item.nplike.full(lenoutindex, -1, np.int64)
-                outindex[~isnegative[expanded]] = item.nplike.arange(
-                    len(nextcontent), dtype=np.int64
-                )
-
-                return ak._v2.contents.IndexedOptionArray(
-                    ak._v2.index.Index(outindex),
-                    ak._v2.contents.NumpyArray(nextcontent, nplike=item.nplike),
-                )
-
-            else:
-                return ak._v2.contents.IndexedOptionArray(
-                    item.index,
-                    item.content,
-                    identifier=item.identifier,
-                    parameters=item.parameters,
-                )
-
-        elif isinstance(item, ak._v2.contents.NumpyArray):
-            assert item.shape == (len(item),)
-            return item
-
-        else:
-            raise AssertionError(type(item))
-
-    def _getitem_next_array_wrap(self, outcontent, shape):
-        length = shape[-2] if len(shape) >= 2 else 0
-        out = ak._v2.contents.RegularArray(outcontent, shape[-1], length, None, None)
-
-        for i in range(len(shape) - 2, -1, -1):
-            length = shape[i - 1] if i > 0 else 0
-            out = ak._v2.contents.RegularArray(out, shape[i], length, None, None)
-        return out
 
     def _carry_asrange(self, carry):
         assert isinstance(carry, ak._v2.index.Index)
@@ -877,29 +512,3 @@ at inner {2} of length {3}, using sub-slice {4}.{5}""".format(
     @property
     def keys(self):
         return self.Form.keys.__get__(self)
-
-
-class NestedIndexError(IndexError):
-    def __init__(self, array, slicer, details=None):
-        self._array = array
-        self._slicer = slicer
-        self._details = details
-
-    @property
-    def array(self):
-        return self._array
-
-    @property
-    def slicer(self):
-        return self._slicer
-
-    @property
-    def details(self):
-        return self._details
-
-    def __str__(self):
-        return "cannot slice {0} with {1}{2}".format(
-            type(self._array).__name__,
-            repr(self._slicer),
-            "" if self._details is None else ": " + self._details,
-        )
