+++ src/awkward/_v2/contents/emptyarray.py
@@ -3,7 +3,8 @@
 from __future__ import absolute_import
 
 import awkward as ak
-from awkward._v2.contents.content import Content, NestedIndexError
+from awkward._v2._slicing import NestedIndexError
+from awkward._v2.contents.content import Content
 from awkward._v2.forms.emptyform import EmptyForm
 
 np = ak.nplike.NumpyMetadata.instance()
