+++ src/awkward/_v2/contents/recordarray.py
@@ -11,7 +11,8 @@ import numpy as np
 
 import awkward as ak
 from awkward._v2.record import Record
-from awkward._v2.contents.content import Content, NestedIndexError
+from awkward._v2._slicing import NestedIndexError
+from awkward._v2.contents.content import Content
 from awkward._v2.forms.recordform import RecordForm
 
 
@@ -244,7 +245,7 @@ class RecordArray(Content):
             return self.content(where)
 
         else:
-            nexthead, nexttail = self._headtail(only_fields)
+            nexthead, nexttail = ak._v2._slicing.headtail(only_fields)
             if ak.util.isstr(nexthead):
                 return self.content(where)._getitem_field(nexthead, nexttail)
             else:
@@ -260,7 +261,7 @@ class RecordArray(Content):
         if len(only_fields) == 0:
             contents = [self.content(i) for i in indexes]
         else:
-            nexthead, nexttail = self._headtail(only_fields)
+            nexthead, nexttail = ak._v2._slicing.headtail(only_fields)
             if ak._util.isstr(nexthead):
                 contents = [
                     self.content(i)._getitem_field(nexthead, nexttail) for i in indexes
@@ -337,7 +338,7 @@ class RecordArray(Content):
             raise self._getitem_next_missing(head, tail, advanced)
 
         else:
-            nexthead, nexttail = self._headtail(tail)
+            nexthead, nexttail = ak._v2._slicing.headtail(tail)
 
             contents = []
             for i in range(len(self._contents)):
